[feat] PR Review - 팀장이 자신을 프로젝트에서 제거할 수 없도록 수정

+++ src/main/java/dev/woori/wooriLog/domain/project/service/ProjectService.java
@@ -167,6 +167,9 @@ public class ProjectService {
     @Transactional
     public void deleteProjectMember(Long projectId, Long leaderId, Long memberId) {
         checkAccessPermission(projectId, leaderId);
+        if (memberId.equals(leaderId)) {
+            throw new IllegalArgumentException(LEADER_CAN_NOT_DELETE);
+        }
         ProjectMember projectMember = projectMemberRepository.findByProject_IdAndMember_Id(projectId, memberId)
                 .orElseThrow(() -> new EntityNotFoundException(RELATION_NOT_FOUND));
         projectMemberRepository.delete(projectMember);
